requirements.txt
mido
pillow